requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.7
django-bootstrap3==7.0.1
django-extensions==1.6.7
djangorestframework==3.3.3
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
